Split TestBaseSwLegacy/Otlp
